+++ src/test-integration/resources/data/javascript.expected/javascript-colon.js
@@ -2,24 +2,18 @@ if (! _$jscoverage['javascript-colon.js']) {
   _$jscoverage['javascript-colon.js'] = {};
   _$jscoverage['javascript-colon.js'].lineData = [];
   _$jscoverage['javascript-colon.js'].lineData[1] = 0;
-  _$jscoverage['javascript-colon.js'].lineData[2] = 0;
   _$jscoverage['javascript-colon.js'].lineData[4] = 0;
   _$jscoverage['javascript-colon.js'].lineData[5] = 0;
   _$jscoverage['javascript-colon.js'].lineData[6] = 0;
 }
+_$jscoverage['javascript-colon.js'].lineData[1]++;
 x:
-  {
-    _$jscoverage['javascript-colon.js'].lineData[1]++;
-    _$jscoverage['javascript-colon.js'].lineData[2]++;
-    y = 0;
-  }
+  y = 0;
+_$jscoverage['javascript-colon.js'].lineData[4]++;
 y:
   {
-    _$jscoverage['javascript-colon.js'].lineData[4]++;
-    {
-      _$jscoverage['javascript-colon.js'].lineData[5]++;
-      let y = 1;
-      _$jscoverage['javascript-colon.js'].lineData[6]++;
-      print(y);
-    }
+    _$jscoverage['javascript-colon.js'].lineData[5]++;
+    let y = 1;
+    _$jscoverage['javascript-colon.js'].lineData[6]++;
+    print(y);
   }